requirements.txt
videosdk-agents
videosdk-plugins-openai
asyncio
scipy
python-dotenv
twilio